feat(supervise): refuse launch when the cap can't fit one iterate
Also writes a live budget file for the session's duration (cleared on
completion or a termination signal) so other commands can read
remaining time, threads session config through a new SupervisedSession
context instead of a growing keyword-argument list, and switches the
wall-clock cap from a single asyncio.sleep to polling real time
against the deadline so it still fires on time after the machine
sleeps mid-run.

+++ src/gymrat/session/budget.py
@@ -38,7 +38,7 @@ class Budget:
     """
 
     started_at_ms: float
-    max_minutes: int
+    max_minutes: float
     deadline_ms: float
     version: int = _BUDGET_VERSION
